traing_eval/requirements.txt
absl-py==2.1.0
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
click==8.1.7
comm==0.2.2
datasets==2.18.0
debugpy==1.8.1
decorator==5.1.1
deepspeed==0.14.0
defusedxml==0.7.1
dill==0.3.8
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
fastobo==0.12.3
filelock==3.9.0
fonttools==4.25.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
grpcio==1.62.1
h5py==3.11.0
hjson==3.1.0
huggingface-hub==0.22.2
ipykernel==6.29.4
ipython==8.23.0
ipython-genutils==0.2.0
jedi==0.19.1
Jinja2==3.1.3
jiwer==3.0.3
joblib==1.4.0
json5==0.9.25
jsonpointer==2.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-server==1.24.0
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab==3.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
lark==1.1.9
lightning==2.2.1
lightning-utilities==0.11.2
llvmlite==0.42.0
lxml==5.2.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
munkres==1.1.4
nbclassic==0.5.6
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
ply==3.11
polars==0.20.21
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
pynvml==11.5.0
PyQt5==5.15.10
pyteomics==4.7.1
pytorch-lightning==2.2.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.8.1
referencing==0.34.0
rpds-py==0.18.0
scikit-learn==1.4.2
seaborn==0.13.2
Send2Trash==1.8.3
sniffio==1.3.1
soupsieve==2.5
spectrum-utils==0.4.1
stack-data==0.6.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.4.0
tinycss2==1.2.1
torch==2.2.1+cu118
torchmetrics==1.3.2
torchvision==0.17.1+cu118
traitlets==5.14.2
triton==2.2.0
typing==3.7.4.3
tzdata==2024.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
xxhash==3.4.1
yarl==1.9.4
